requirements.txt
filetype==1.2.0
GitPython==3.1.30
pathlib2==2.3.7.post1
python-telegram-bot==20.0
requests==2.28.2
translators==5.5.6
urllib3==1.26.14
urwid==2.1.2
wakeonlan==3.0.0
websockets==10.4
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz